Project: Classic Arcade Game Clone
This is from Udacity’s FEND Online Course, Project 3. To complete the
project, I had to use the provided tools and info to create a clone of
the classic arcade game often referred to as “Frogger”. A game board,
moveable player and randomly moving enemies were added to the game.